chore: moved pagination methods to own module
