+++ iris-client-fe/src/views/index-tracking-list/index-tracking-list.view.vue
@@ -88,7 +88,7 @@ import {
   getPageSizeFromRouteWithDefault,
   getStatusFilterFromRoute,
   getStringParamFromRouteWithOptionalFallback,
-} from "@/utils/misc";
+} from "@/utils/pagination";
 
 function getFormattedDate(date?: string): string {
   return date
